V2.3 Released
!nick @user nickname

If user already has a nickname -> do you want to add "nickname" as a new nickname, current nicknames for user are: nick1, nick2, etc.
Have bot then react to that message with a check mark and x emoji, whichever the user presses is the decision to either add the new nickname or not
make sure only the use who sent the original !nick command is able to do this

This double check functionality might be annoying if you have to do it over and over for 100s of users, so this should also be able to be toggled off with nick_doublecheck in the config which I have already added.

ensure nickdelete works as intended

Add a double check in the same way that !nick has one as this will remove all of the dkp data for a user's toon.
Double check message: "Are you sure you want to remove the nickname "nickname" from User? This will clear all dkp data from "nickname". This cannot be reversed."
Have bot then react to that message with a check mark and x emoji, whichever the user presses is the decision to either add the new nickname or not
make sure only the use who sent the original !nick command is able to do this

This should also be toggleable with nickdelete_doublecheck in the config.

add !setmain @user nickname
reminder: there is a config ```["role_req_setmain", "true"]``` that needs to be checked before moving forward with the command much like how we did it with all other commands.
If main isn't set treat them like normal and use the @user line in the Balances_Database.csv file.

add the main and nickname support to !k and !a (!a first because easier)

time to fix nicknames again, granting things to any nickname always grants to the "main" even when specifying a different nickname, this obviously should not be the case. Only when doing @user should the "main" nickname be substituted in

make it so that all dkp commands (!a, add, subtract, etc can send messages (toggleable) in the dkp keeping log (maybe even include option to have the person who granted the dkp show up in the log))

Add toggles to pick which commands send messages in the #dkp-keeping-log channel

make commands to do those toggles

fix the !k functionality (if no nickname etc) also make the !k message does not just post in the general chat
make sure that the !k command is able to have message toggles

Add command that is able to operate all toggles easier more like the editrolecommands

add functionality so that doing !setmain and not specifying a user will assume that the user is you

add command that is able to more easily do help stuff more like the editrolecommands 
done 1/22/2025
add auction message toggles for dkp keeping channel and make sure that they are in the !toggles command
update all auction commands

add more functionality to !auctionend to be able to display a list of auctions so you can pick which to end
do same with !auctioncancel

Update !bal command to cope with alts, if !bal is done and they only have 1 nickname & that nickname is set as main then only display the balance of that nickname. If they have 1 nickname (& the nickname is not set as main) then the bot should send a popup (only able to be pressed by the person who did the !bal command) allowing the user to make a selection between their username and the nickname. If the user has multiple nicknames and 1 is set as their main, then it should only show the nicknames and not their username as options. Just to make it clear: Whenever a nickname is set to the main nickname it should take the place of the user's account.

add !setmain {nickname} usage

currently !removemain doesn't work correctly
add !removemain {no nickname} usage

add !removemain help page

update the !nick and !nickdelete help pages (add mentions to the dkpaddcurrent and dkpsubtractlifetime commands too)
add !setmain help page

Add Help menu that has a youtube playlist link thatI can add how-to pages to.

On start make sure that the config.csv file has all of the correct settings, and if it is missing a setting have that auto-populate with the default

add toggles for Active_timers dkp_vals_channel nick_doublecheck nickdelete_doublecheck


Make it so the toggles that toggle on or off channels or rolls will also call functions to remove / add those channels / roles.
Toggles that need to be updated when !toggles buttons are used:
Have toggle_dl, toggle_edl, toggle_legacy, and toggle_worldboss when toggled true create the text channel like the command does, and when toggled false the text channel is removed
Have toggle_dl_role, toggle_edl_role, toggle_legacy_role, and toggle_worldboss_role when toggled true create their corresponding roles, when toggled false the roles are deleted

Have toggle_role_channel when toggled true create the channel, and when toggled false delete the channel in the same way that the !togglerolechannel does
Have toggle roles also change the embed in !get-timer-roles channel if it exists

have toggling dkp_vals_channel true create the channel, and when toggled false delete the channel in the same way that the !toggledkpchannel command does

fix it so the roles toggled update embed again as it has broken

fix it so it doesn't update the config every time previous or next buttons are pressed in the !toggles command

add toggle for timers to the !toggles command and ensure it works correctly

write help page for !toggles

fix toggle_role help button

make it so !cancel displays an embed with all active timers allowing for easier cancelation
update !cancel help page